+++ fidget/benches/render.rs
@@ -26,10 +26,7 @@ pub fn prospero_size_sweep(c: &mut Criterion) {
         group.bench_function(BenchmarkId::new("vm", size), move |b| {
             b.iter(|| {
                 let tape = shape_vm.clone();
-                black_box(fidget::render::render2d::<
-                    _,
-                    fidget::render::BitRenderMode,
-                >(tape, cfg))
+                black_box(cfg.run::<_, fidget::render::BitRenderMode>(tape))
             })
         });
 
@@ -43,10 +40,7 @@ pub fn prospero_size_sweep(c: &mut Criterion) {
             group.bench_function(BenchmarkId::new("jit", size), move |b| {
                 b.iter(|| {
                     let tape = shape_jit.clone();
-                    black_box(fidget::render::render2d::<
-                        _,
-                        fidget::render::BitRenderMode,
-                    >(tape, cfg))
+                    black_box(cfg.run::<_, fidget::render::BitRenderMode>(tape))
                 })
             });
         }
@@ -74,10 +68,7 @@ pub fn prospero_thread_sweep(c: &mut Criterion) {
         group.bench_function(BenchmarkId::new("vm", threads), move |b| {
             b.iter(|| {
                 let tape = shape_vm.clone();
-                black_box(fidget::render::render2d::<
-                    _,
-                    fidget::render::BitRenderMode,
-                >(tape, cfg))
+                black_box(cfg.run::<_, fidget::render::BitRenderMode>(tape))
             })
         });
         #[cfg(feature = "jit")]
@@ -91,10 +82,7 @@ pub fn prospero_thread_sweep(c: &mut Criterion) {
             group.bench_function(BenchmarkId::new("jit", threads), move |b| {
                 b.iter(|| {
                     let tape = shape_jit.clone();
-                    black_box(fidget::render::render2d::<
-                        _,
-                        fidget::render::BitRenderMode,
-                    >(tape, cfg))
+                    black_box(cfg.run::<_, fidget::render::BitRenderMode>(tape))
                 })
             });
         }